+++ src/main/java/com/windfall/api/review/dto/response/ReviewDetailsRaw.java
@@ -1,6 +1,7 @@
 package com.windfall.api.review.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.time.LocalDateTime;
 
 public record ReviewDetailsRaw(
 
@@ -20,7 +21,7 @@ public record ReviewDetailsRaw(
     String nickname,
 
     @Schema(description = "별점")
-    int rating,
+    double rating,
 
     @Schema(description = "내용")
     String content
